Troquei a estilização de Links do component ProfileStyles para o SharedStyles
