+++ src/styles/SharedFiles.jsx
@@ -1,5 +1,6 @@
 import { styled } from 'styled-components';
-
+import { Link } from "react-router-dom";
+import { BsFillHouseFill } from "react-icons/bs";
 
 // Estilizações das Rotas de Camisas
 export const Container = styled.div`
@@ -24,4 +25,35 @@ export const MapWrapper = styled.div``
 export const Title = styled.h1`
   font-size: 34px;
   margin: 0 0 20px 0;
+`
+
+//LINKS
+export const ProfileLinksContainer = styled.div`
+    width: 80%;
+    margin: 20px auto;
+    display: flex;
+    align-items: center;
+    gap: 0 8px;
+`
+export const ProfileHomeLink = styled(BsFillHouseFill)`
+    font-size: 17px;
+`
+export const ProfileSlash = styled.span`
+    font-family: "roboto"; 
+    font-size: 18px;
+    &:hover {
+        transition: 0.3s;
+        cursor: pointer;
+        color: #f82e56;
+    }
+`
+export const ProfileLink = styled(Link)`
+    color: black; 
+    font-family: "roboto";
+
+    &:hover {
+        transition: 0.3s;
+        cursor: pointer;
+        color: #f82e56;
+    }
 `
